# src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tabibi API",
    "version": "1.0.0",
    "description": "Healthcare CRM API - Backend Next (Bun/Express/Prisma/PostgreSQL)"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/appointments": {
      "get": {
        "summary": "List appointments",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of appointments"
          }
        }
      },
      "post": {
        "summary": "Create appointment",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "201": {
            "description": "Appointment created"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}